First, you will need to create a board for the game. You can do this using a 2D list or a dictionary.
Next, you will need to create a function that allows the user to make a move. This function should take in the current board and the position where the user wants to place their symbol (X or O).

You will also need to create a function that checks if the game has been won or tied. This function should check the rows, columns, and diagonals of the board to see if there is a winner.

Create a function to handle the computer's moves. This function should use some strategy to choose the best move based on the current state of the board.

You will also need a main game loop that handles alternating turns between the user and the computer, and continues until the game is won or tied.

You may also want to add additional features like the ability to play again or choose the symbol (X or O) for the user.

Finally, you can add some visual elements using a library like Pygame to make the game more interactive and visually appealing.